Move test-only functions off TradingStateMachine
