+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/IsolatedMarginModeTests.kt
@@ -12,6 +12,7 @@ import exchange.dydx.abacus.state.model.closePosition
 import exchange.dydx.abacus.state.model.trade
 import exchange.dydx.abacus.state.model.tradeInMarket
 import exchange.dydx.abacus.tests.extensions.parseOnChainEquityTiers
+import exchange.dydx.abacus.tests.extensions.socket
 import kotlin.test.BeforeTest
 import kotlin.test.DefaultAsserter.assertTrue
 import kotlin.test.Test
