+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/TriggerOrdersInputTests.kt
@@ -7,6 +7,7 @@ import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.app.adaptors.AbUrl
 import exchange.dydx.abacus.state.model.TriggerOrdersInputField
 import exchange.dydx.abacus.state.model.triggerOrders
+import exchange.dydx.abacus.tests.extensions.rest
 import exchange.dydx.abacus.utils.Rounder
 import kotlin.math.abs
 import kotlin.test.Test
