+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/v4/V4AccountOrdersSortingTest.kt
@@ -3,6 +3,7 @@ package exchange.dydx.abacus.payload.v4
 import exchange.dydx.abacus.calculator.CalculationPeriod
 import exchange.dydx.abacus.tests.extensions.loadv4SubaccountsWithPositions
 import exchange.dydx.abacus.tests.extensions.log
+import exchange.dydx.abacus.tests.extensions.socket
 import exchange.dydx.abacus.utils.SHORT_TERM_ORDER_DURATION
 import exchange.dydx.abacus.utils.ServerTime
 import kotlin.test.Test
