+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/v4/V4DuplicateWebsocketMessageTests.kt
@@ -2,6 +2,7 @@ package exchange.dydx.abacus.payload.v4
 
 import exchange.dydx.abacus.calculator.CalculationPeriod
 import exchange.dydx.abacus.tests.extensions.loadv4TradesChanged
+import exchange.dydx.abacus.tests.extensions.socket
 import kotlin.test.Test
 import kotlin.test.assertEquals
