+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/v4/V4OrderbookTests.kt
@@ -4,6 +4,7 @@ import exchange.dydx.abacus.payload.v3.V3BaseTests
 import exchange.dydx.abacus.state.app.adaptors.AbUrl
 import exchange.dydx.abacus.state.model.setOrderbookGrouping
 import exchange.dydx.abacus.tests.extensions.log
+import exchange.dydx.abacus.tests.extensions.socket
 import exchange.dydx.abacus.utils.ServerTime
 import kotlin.test.Test
 import kotlin.test.assertEquals
