+++ src/commonTest/kotlin/exchange.dydx.abacus/tickets/TRCL2998Tests.kt
@@ -4,6 +4,7 @@ import exchange.dydx.abacus.calculator.CalculationPeriod
 import exchange.dydx.abacus.payload.v4.V4BaseTests
 import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.model.tradeInMarket
+import exchange.dydx.abacus.tests.extensions.socket
 import kotlin.test.Test
 import kotlin.test.assertEquals
