+++ src/commonTest/kotlin/exchange.dydx.abacus/validation/TriggerOrdersInputValidationTests.kt
@@ -8,6 +8,7 @@ import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.app.adaptors.AbUrl
 import exchange.dydx.abacus.state.model.TriggerOrdersInputField
 import exchange.dydx.abacus.state.model.triggerOrders
+import exchange.dydx.abacus.tests.extensions.rest
 import kollections.iListOf
 import kotlin.test.Test
 import kotlin.test.assertEquals
