+++ src/commonTest/kotlin/exchange.dydx.abacus/validation/ValidationsTests.kt
@@ -4,6 +4,7 @@ import exchange.dydx.abacus.payload.v4.V4BaseTests
 import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.model.trade
 import exchange.dydx.abacus.tests.extensions.parseOnChainEquityTiers
+import exchange.dydx.abacus.tests.extensions.socket
 
 open class ValidationsTests : V4BaseTests() {
     override fun setup() {
